Handled distinctive NFDBits and func index's return values according to different architectures.
